feat: add "dirty" fixes
Problem:
On a huge codebase, it's a lot of work to find and fix the unqualified usages
that this rule guards against.

Solution:
Supply a Fix that will qualify the usages.
Note that **the fix may have false-positives,** and add incorrect qualifications to
functions or values that coincidentally have the same name as an imported function.
Hence the "dirty" in the title.

Even then, implementing this dirty Fix and cleaning up the resulting errors was
**much** faster than any alternative I could think of, on our 160k+ LoC project.